backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
psycopg2-binary>=2.9.0
